Move most of chatlas logic to `_chat_bookmark.py`
